refactor : move mdx component to article folder
