+++ pages/programming/[id].page.tsx
@@ -14,9 +14,9 @@ import { detailPageState } from 'src/recoil/article';
 import { ViewArticleElement } from 'src/types/article';
 import { MDXRemote } from 'next-mdx-remote';
 
-import MDXDetail from 'src/components/mdx/MDXDetail';
-import MDXTitle from 'src/components/mdx/MDXTitle';
-import MDXTag from 'src/components/mdx/MDXComponent';
+import MDXTag from 'src/components/article/article-detail/MDXTag';
+import ArticleDetail from 'src/components/article/article-detail/ArticleDetail';
+import ArticleDetailTitle from 'src/components/article/article-detail/ArticleDetailTitle/ArticleDetailTitle';
 import Head from 'src/components/Head';
 import Image from 'next/image';
 
@@ -60,8 +60,8 @@ const Article = ({
   return (
     <>
       <Head article={article} />
-      <MDXDetail
-        mdxTitle={<MDXTitle article={article} />}
+      <ArticleDetail
+        mdxTitle={<ArticleDetailTitle article={article} />}
         image={
           <Image
             src={article.imgUrl}
